Exception global

+++ src/main/java/com/dh/DentalClinicMVC/service/impl/DentistServiceImpl.java
@@ -1,6 +1,7 @@
 package com.dh.DentalClinicMVC.service.impl;
 
 import com.dh.DentalClinicMVC.entity.Dentist;
+import com.dh.DentalClinicMVC.exception.ResourceNotFoundException;
 import com.dh.DentalClinicMVC.repository.IDentistRepository;
 import com.dh.DentalClinicMVC.service.IDentistService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,8 +35,15 @@ public class DentistServiceImpl implements IDentistService {
     }
 
     @Override
-    public void delete(Long id) {
-        dentistRepository.deleteById(id);
+    public void delete(Long id) throws ResourceNotFoundException {
+        //Buscamos al odontologo por id
+        Optional<Dentist> dentistToLookFor = findById(id);
+        if (dentistToLookFor.isPresent()) {
+            dentistRepository.deleteById(id);
+        } else{
+            throw new ResourceNotFoundException("No se pudo eliminar el odontologo con el id: " + id);
+
+        }
     }
 
     @Override
